+++ docs/release-notes.md
@@ -24,7 +24,7 @@
 - Add embedded path implicit to select_related paths.
 - `crawl_relationship` has now a better error when a field does not exist.
 - Unpacking and accessing `CrawlResult` (the result of `crawl_relationship`) is now deprecated. Use the results attributes instead.
-- Don't select unneccessary intermediate columns when specifying select paths. This is potential breaking.
+- `select_related` has now a sparse mode, which will become the default in future. It doesn't select all columns on the intermediate paths but only the ones required for traversal.
 
 ### Fixed
 
@@ -41,8 +41,10 @@
 ### Breaking
 
 - `crawl_relationship` returns now not `exact` as `operator`, when no operator was found but an empty string.
-- `select_related` intermediate paths aren't fully loaded anymore. Only the relevant columns for traversing are selected.
-  Upgrade `select_related("company__user")` to `select_related("company","company__user")`
+- `select_related` intermediate paths will not be fully selected anymore by default in future.
+  In the new sparse mode only the relevant columns for traversing are selected. It will become the default in future.
+  A DeprecationWarning will be issued if you are affected.
+  Upgrade `select_related("company__user")` to `select_related("company","company__user")` or use the keyword `sparse=False`.
 
 ## 0.36.1